Move new stuff to new files
